Allow connecting to an existing shiny instance by specifying the URL.

+++ R/initialize.R
@@ -4,7 +4,7 @@
 
 sd_initialize <- function(self, private, path, loadTimeout, checkNames,
                           debug, phantomTimeout, seed, cleanLogs,
-                          shinyOptions) {
+                          shinyOptions, url=NULL) {
 
   private$cleanLogs <- cleanLogs
   if (is.null(loadTimeout)) {
@@ -21,9 +21,13 @@ sd_initialize <- function(self, private, path, loadTimeout, checkNames,
   self$logEvent("Getting PhantomJS port")
   private$phantomPort <- get_phantomPort(timeout = phantomTimeout)
 
-  if (grepl("^http(s?)://", path)) {
-    private$setShinyUrl(path)
+  if(!dir_exists(path))
+    stop("No such directory: ", path)
+  private$path = path
 
+  if(!is.null(url)) {
+    "!DEBUG setting URL provided by the user"
+    private$setShinyUrl(url)
   } else {
     "!DEBUG starting shiny app from path"
     self$logEvent("Starting Shiny app")
@@ -247,7 +251,7 @@ find_phantom <- function(quiet = FALSE) {
 sd_finalize <- function(self, private) {
   self$stop()
 
-  if (isTRUE(private$cleanLogs)) {
+  if (!is.null(private$shinyProcess) && isTRUE(private$cleanLogs)) {
     unlink(private$shinyProcess$get_output_file())
     unlink(private$shinyProcess$get_error_file())
   }
